Address more review feedback
